Rename package from woopra to amplitude
