+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
 quarkus.swagger-ui.always-include=true
 quarkus.ssl.native=true
-quarkus.arc.selected-alternatives=com.redhat.devworkspace.services.telemetry.woopra.MainConfiguration,com.redhat.devworkspace.services.telemetry.woopra.AnalyticsManager
+quarkus.arc.selected-alternatives=com.redhat.devworkspace.services.telemetry.amplitude.MainConfiguration,com.redhat.devworkspace.services.telemetry.amplitude.AnalyticsManager
